fix a matching problem

+++ contrib/brl/b3p/netlib/b3p_netlib.h
@@ -11,6 +11,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
+
 int trapru_(double f(float*), float *a, float *b, int *m, float *	trule);
 
 int simpru_(double f(float* ), float *a, float *b, int *m, float *srule);
@@ -26,12 +27,11 @@ int simpru_(double f(float* ), float *a, float *b, int *m, float *srule);
 //  errbdd: error estimation
 //  m: substates
 */
-int adaptquad_(int f(float*), float *a, float *b, float *tol, float * srmat, float *integral, float *errbdd, int *m, int *state);
+int adaptquad_(double f(float*), float *a, float *b, float *tol, float * srmat, float *integral, float *errbdd, int *m, int *state);
 
 #ifdef __cplusplus
 }
 #endif
-
 
 #endif
